Refactor telemetry MSP.
BOXNAMES is repeated in case the first transmission is never received.

Each time sendMspTelemetry is called it executes one and only one
command.  There was nothing wrong with the old approach which sometimes
sent more than one command each time sendMspTelemetry was called but
this is simpler and there is far less duplication.