Update battery warning and turn indicators to use led config.
Cleaned up the battery warning checking as it was being calculated twice
instead of just once per battery warning state cycle.

Updated the code so that modes are shown when disarmed.  Any led's
without the MODE flag will then show the armed state.  In the config in
this commit that results in the corner LEDs being GREEN when disarmed
and OFF when armed.