Update led config so it's possible to show which LEDs show the arm state.  Mutually exclusive with leds that show MODE.
Use blue to indicate armed, green for disarmed.

Leds not configured for ARM_STATE or MODE will be off by default.

The example combines corner LEDs with ARM_STATE and INDICATOR.  This
results in the corner leds showing arm state and then being overridden
with an indicator color as needed.