Ported OpenPilot PPM frame validation.
On the CC3D it was observed that jitters were present, this code
validates each frame.

Prior to this, on other boards (Naze32/Flip32+/Olimexino/etc) it was
observed that when using an 8 channel RX AUX5-8 would sometimes be set
due to bad PPM data.

This code fixes both issues.