Fix WS2812 LED initialisation.
The timer was not being initialised correctly, it only worked by
accident due to timer configuration of other timer channels.  This
explained why TIM3 didn't work in the mapping exclusion for the CC3D and
why for the LEDs to work correctly TIM2 had to be specified so that TIM3
was actually accidentally initialised still.