First attempt at getting CC3D PPM to work when oneshot is enabled.
I don't think that this will work, as we are manipulating the timer value to trick it into overflowing.  Better would be to force an overflow a different way, that would preserve the current timer value.  This would then give the proper value to the CC3D PPM routine, provided it was then divided by 8.