Update Buzzer.md
I added information to help anyone who runs into the same problem I did, the buzzer not functioning after trying multiple devices.  I researched and saw you need to have 5v power to the board. The ESC connection is a simple place to power the board when working on the flight controller on a bench. This allows you to use the buzzer for conformations.  

Please see added text and link to a Radio Shack (local US electronics retailer) buzzer.