FrSky GPS code size and de-duplication improvements.
I noticed this after merging 593db0ce1eb47e8ddc95ea7fab03c37871818cbb.

If the telemetry screen shows no fix when the fix is lost then perhaps
the intention was to make localGPS_coord static in the original code,
but if GPS_coord retains it's value when a fix is lost then this is ok,
please test!