config: add a 'format' field; switch to a terminated registry.
Add a (hopefully never used) 'format' field that can be used to
invalidate the PG when the meaning is changed.

Use a null entry to terminate the registry array instead of a symbol.
This should make the code testable as it's easy to mock an array but
hard to mock the linker.

Signed-off-by: Michael Hope <mlhx@google.com>