Add mention of power requirements for gpassthrough
This change adds a brief mention of power requirements and USB power pass-through to the GPS docks. This small but important fact is not mentioned anywhere, though it is obvious if you stare at the board specific documentation long enough. Many larger flight controllers (eg Pixhawk) do pass through 5V from USB, so this is non-obvious to users of those FCs.