Working - except when exiting from bootloader
Fixes the `Waiting for Data` problem in configurator when:

- Clicking on `Save and Reboot` button
- Clicking on `Flash Firmware`  button

When the system reboots after programming the new firmware it still
needs a connect/disconnect cycle to work normal. This cannot be solved
since the bootloader is a fixed program in ROM.

Fix is for F1 targets and only tested on naze boards.