Update config.c
Not that important, but may help common issues and questions that pertain only to Betaflight blackbox use. 
Change default to set blackbox_rate_denom = 2 since Betaflight runs at 1000 looptime and a 1/1 can causes corruption at that speed to the blackbox log. 1/2 will work for everyone and 1/1 can be configured later if desired. Spoke to Joshua Bardwell and he is using 1/1 but must disable ACC for it to be stable. Blackbox.md explains the use and benefit to a 1/1 setting so a user is likely to try that too, but a default 1/2 will at least get it working to start.