Fix failure to recover from badly received UBLOX GPS data.
The problem occurs when the payload size was determined to be 65535.
The bug was due to overflowing the payload counter before checking it
against the size.

This also highlights that it takes along time to recover from a badly
received packet as the parser waits for data to come in.  Likely a check
should be made that checks for invalid packet lengths;  Another commit
to follow.