Recover quicker from a badly received / incorrect UBLOX payload size.
Previously the code would wait for additional data to be received, which
could have actually contained valid UBLOX packets, in the case where one
or two bytes were incorrectly received during payload size reception.

Now the code just starts looking for the start of the a new UBLOX packet
if the payload size is greater than the largest known UBOX payload we
expect to handle.