decouple configuration storage from MSP protocol.
While using the MSP id's for the parameter group ID's could be used and
might save a little space I don't think we should tie the the
configuration storage implementation to the MSP protocol.

This change still allows MSP commands that use the exact same data to
work.  It does this by finding appropriate MSP command ids for a
parameter groups from a table using a callback mechanism.

The API for parameter groups now no-longer has a depencency on MSP.
Failsafe no longer has a dependency on MSP.
Board alignment no longer has a dependency on MSP.

This change also bumps the size of the ID from a unit8_t to a unit16_t -
i didn't feel that 255 would be enough and it's not something we want to
deal with chaning in the future when we run out of IDs in the 8bit
space.