Migrate gimbal configuration (per-profile) use parameter group configuration.
See config_eeprom_unittest.cc for an example.

The basic idea is that we still hold data for all three profile in ram
all the time.  The flash page size prevents us from doing otherwise
(efficiently).

Each PG registation is annotated with flags to define wether it is a
system or a profile registration.

For profile registrations there are three main points:
1) the 'base' field points to an area of memory big enough to hold 3
copies of the configuration.
2) a 'ptr' field is set to an address of a pointer that will be updated
to point to one of the copies.
3) when a profile is activated the pointer is updated to point to the
appropriate configuration.