extract rcControlsConfig from profile_t.  extract flight3DConfig from master_t and rename to motor3DConfig.  Move throttle3d_deadband from flight3DConfig/motor3DConfig to rcControlsConfig_t where it really belongs.  Update MSP_3D and MSP_DEADBAND appropriately.  Bump MSP version.
The issue was that 3 settings in flight3DConfig were for motor output,
used by the mixer to control the ESCs, and 1 was for rx input, used for
checking stick position and throttle state.

Moving them all in one go meant adding inappropriate dependencies to
failsafe.c so they were split.

the 3d throttle code in mixTable() is much more readable now.

Note: This will need a corresponding configurator change.