First cut of Cleanflight OSD support.
The main point is that there is loads of great re-usable code for both
OSD and FCs and allow the both to be configured using the Cleanflight
configurator.

The Cleanflight codebase can now be used to build TWO types of targets:
FC or OSD.

This commit seperates as all the FC specific code into the src/main/fc
directory, OSD code goes in src/main/osd.  Shared code is everywhere
else.  some files have an `_fc` suffux to allow the files to be
distinguished (e.g. on IDE editor tabs).

The idea is that we can now build F1 or F3 based OSDs that use the
MAX7456 IC.  The display logic should be reusable for other OSD driver
ICs (e.g. BrainFPV).  The OSD code should also be usable if the MAX7456
is directly connected to an FC in the future.  We'll cross that bridge
when we come to it however as right now the scheduler tasks, bootup
sequence and other code will likely have to be refactored/reoganized
somewhat.  Later commits will add OSD code to the STM32F3DISCOVERY
target and perhaps an an STM32F3DISCOVERYOSD for development of OSD only
boards.