Little optimisation PID.c
Hi,
I know, it's not a huge optimisation but I just switched the condition of the IF condition (line 425). Given that the YAW axis occur one time in the FOR loop (line 366), the IF(axis == YAW) condition will be called one time whereas the IF(axis != YAW) condition will be called two times.

Check first the most used condition in an IF condition is a good thing to do in a code source.

I just don't know if it will be a major improvement, I let you judge.