Updated after review.
The review was not exhaustive so it's hard to see if there really was
concensus or not.
I have used my own judgement from the comments received but please tell
if something more should be removed or changed.

I would like to add something about indenting switch/case. I think
Astyle does it right but a lot of the current code use extra indents.

I would like to add something about comments. Along the lines of "less
is more but zero is bad".

After this, someone should update docs/development/Development.md too.
Some statements in that is now in this document and one place should be
enough.
At least, the old reference to the baseflight coding style should be
replaced.