Use either virtual or adc amperage meter.
It's possible to use them both at the same time but there are a few
issues to fix before that can happen.

* telemetry needs amperage
* msp needs amperage for OSDs
* blackbox need amperage

There would have to be a way to set the source for each thing and that
would likely lead to confusion and complication that most users do need.

This also saves a few CPU cycles by only calculating what is needed.