Rename 'telemetry_flvss_cells' to 'telemetry_send_cells'.
The former was too specific to be in telemetryConfig_t since it is frsky
and smartport specific.

I considered moving it to frskyConfig_t but then smartport.c would have
a dependency on frsky.c - bad.  another option would be to have
duplicate settings in smartportConfig.c and frskyConfig.c but also bad.

This change seems like an acceptable naming solution which could be
applicable for any telemetry system and doesn't mess up the
dependencies.