Fix mag readings being reset when the mag isn't ready.
The problem here is that some mag's have the ability to be asked if they
have data, but the cleanflight api doesn't leverage this.

Perhaps CF should first ask if the mag has data, and then read it and
reset the values only if the mag read fails.  Even then I'm note sure
that's better than keeping the last heading (e.g. for telemetry when
finding a lost quad)