Remove historical OSD_CALLS_CMS
This was a part of the original code, and was left here so there will
be an option to call CMS code from OSD task so that there is no need
for a separate CMS task.
It was eventually unused; there are couples of important logics around
cmsOpenMenu so doing this will require the same amount of code.