MAX7456: Adjustable screen area position offsets
There are displays and goggles with slightly different timing
characteristics compared the standard ones. With such devices, screen
area may be too high or too low that top or bottom most characters may
be truncated.

This PR provides a way to adjust the screen area position so that
entire screen area can be shifted up or down, or left or right.

A new vcdProfile_t (VCD = Video Character Device) is defined and
video_system was moved from osdProfile also.