fixes issue with power changes without prior frequency change
As trampConfFreq & trampConfPower was not initialised properly with the current values, this would lead to the config not being properly changed (you cannot set the frequency to 0, which is what we tried earlier).

Also, a fixed number of retries has been added to make the code safer and finish after a fixed number of iteration, no matter what.