Fix #2364
USE_I2C was still defined even though no I2C devices BARO MAG were
defined. This led to the I2C being active but not being handled. So any
data on the I2C line caused it to error and not be handled.

Tested with IA6B RX hooked upto the flexport of the CC3D with TX on
before the FC boots.