CF/BF - Fix AK8963 mag broken by 60c2b812d1113818948e9983bd5fb10053970f16.
I think the cause of the broken commit was a search/replace failure
since SPI WRITES were replaced with I2C WRITES to a NULL I2C device.

This is a quick hack to get things working again, no-doubt @martinbudden
will want to inspect the changes and make appropriate changes.