CF/BF - Fix HoTT telemetry.
Two issues: 
1 - failure to work on softserial ports.

The TX pin was not allocated due to the initial port mode.

HoTT telemetry is different in that it changes the serial port mode
between sending and receiving.

This change opens the port in RX/TX mode so that both RX and TX pins are
initally allocated.

Test scenario:
* HoTT on SoftSerial 1
* Diode connecting between RX and TX pins.

2 - bidirectional hardware ports were not supported.

Renamed `sport_halfduplex` to `tlm_halfduplex`.  The setting is now used
by sport and hott telemetry.