CF/BF - Implement PPM to SERVO mixer.
This allows an FC to be used as a decoder board.  The mixer name is a
misnomer as it works for all RX inputs, not just PPM.

Use in conjunction with the resource and servo commands, all RC inputs
are mapped directly to servo outputs.  servo mix rules are NOT applied.