Change airmode activation to be throttle percent based to fix 3D mode
The previous logic used an absolute throttle value of 1350us which won't work in 3D mode and resulted in airmode being activated on arming.  Renamed the parameter to airmode_start_throttle_percent and set the default to 32 (equivalent to what the previous setting of 1350 when min_check is taken into account).

To preserve MSP functionality the value is transformed to/from microsecond values (32 becomes 1320) when interfaced.