Reactivate Runaway Takeoff Prevention after using crash-flip
Reset Runaway Takeoff Prevention so that on the next arming after crash-flip is used it will once again be active.  The use-case is that the pilot may have broken props or other damage that could cause a runaway event.  After successful takeoff the feature will deactivate for the remainder of the battery as before.