Change crash-flip OSD message to track actual state rather than switch
Previously the OSD crash-flip warning would only track the state of the activation switch.  So if the switch was disabled but still armed the message would go away, but the quad was still in active crash-flip mode until it was disarmed.  This change makes the OSD message follow the actual internal state for crash-flip.