Optimize ledstip by disabling unassigned overlays
The current design continuously runs timer loops for all overlays regardless of whether or not any LEDs are actually using the overlay.  Then each overlay function iterates through all configured LED's to apply if needed.

This optimization "looks ahead" and turns off any overlays that are not assigned to any LEDs.  The most optimization is achieved when no overlays are used, and regresses to match the current performance level if every overlay is assigned.