Add accelerometer calibration values to the CLI - also change mag calib representation
Allows the accelerometer calibration to be exported in a dump or diff and restored after upgrading.

The accelerometer calibration is specific to the frame and orientation in which the flight controller is installed.  The users should have a way to restore a correct setting during upgrades without having to recalibrate after every flash.  Replicates the way the magnetometer calibration is exported.

Added acc_calibration as an array type to the CLI.  For consistency changed magnetometer representation from magzero_X, magzero_Y, magzero_Z to be a single array mag_calibration.